+++ class/actor.lua
@@ -117,7 +117,8 @@ methods_actor = {
         local can_pierce = false
         if pierce_multiplier then can_pierce = true end
 
-        local damager = gm._mod_attack_fire_bullet(self.value, x, y, range, direction, damage, hit_sprite or -1, can_pierce, true).attack_info
+        local inst = gm._mod_attack_fire_bullet(self.value, x, y, range, direction, damage, hit_sprite or -1, can_pierce, true)
+        local damager = inst.attack_info
         damager.damage_color = Color.WHITE_ALMOST
 
         -- Set pierce multiplier
@@ -130,25 +131,27 @@ methods_actor = {
             damager.tracer_kind = tracer
         end
 
-        return Damager.wrap(damager)
+        return Damager.wrap(damager), Instance.wrap(inst)
     end,
 
 
     fire_explosion = function(self, x, y, width, height, damage, explosion_sprite, sparks_sprite)
-        local damager = gm._mod_attack_fire_explosion(self.value, x, y, width, height, damage, explosion_sprite or -1, sparks_sprite or -1, true).attack_info
+        local inst = gm._mod_attack_fire_explosion(self.value, x, y, width, height, damage, explosion_sprite or -1, sparks_sprite or -1, true)
+        local damager = inst.attack_info
         damager.damage_color = Color.WHITE_ALMOST
 
-        return Damager.wrap(damager)
+        return Damager.wrap(damager), Instance.wrap(inst)
     end,
 
 
     fire_direct = function(self, target, damage, direction, x, y, hit_sprite)
         target = Wrap.unwrap(target)
 
-        local damager = gm._mod_attack_fire_direct(self.value, target, x or target.x, y or target.y, direction or 0, damage, hit_sprite or -1, true).attack_info
+        local inst = gm._mod_attack_fire_direct(self.value, target, x or target.x, y or target.y, direction or 0, damage, hit_sprite or -1, true)
+        local damager = inst.attack_info
         damager.damage_color = Color.WHITE_ALMOST
         
-        return Damager.wrap(damager)
+        return Damager.wrap(damager), Instance.wrap(inst)
     end,