+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### v1.1.22
 * Monster_Card : Added `SPAWN_TYPE` enum.
+* Actor
+    * `fire_` methods now return the actual attack instance as a second unpacked return value.
 
 ### v1.1.21
 * Damager : Added `add_offset`.